Reducing duplicated code for Cypress tests (#2568)
* Cleaning up imports and whitespace

* Refactoring with testInitializeAudit and testFederalAwards

* Fixing testFileUploadMsg

* Comment

* Import fix

---------
