+++ backend/cypress/e2e/workbook-uploads.cy.js
@@ -1,6 +1,4 @@
 import 'cypress-file-upload';
-//require("@4tw/cypress-drag-drop");
-
 import {
   testWorkbookFederalAwards,
   testWorkbookFindingsUniformGuidance,
