+++ backend/cypress/support/tribal-audit-form.js
@@ -1,16 +1,15 @@
-
 export function testTribalAuditPublic(){
-    cy.get(`#is_tribal_information_authorized_to_be_public-yes`).click({force:true});
-    cy.get('#tribal_authorization_certifying_official_name').type('John Wick');
-    cy.get('#tribal_authorization_certifying_official_title').type('Offical');
-    cy.get('#continue').click();
-    cy.url().should('match', /\/audit\/submission-progress\/[0-9]{4}-[0-9]{2}-GSAFAC-[0-9]{10}$/);
+  cy.get(`#is_tribal_information_authorized_to_be_public-yes`).click({force:true});
+  cy.get('#tribal_authorization_certifying_official_name').type('John Wick');
+  cy.get('#tribal_authorization_certifying_official_title').type('Offical');
+  cy.get('#continue').click();
+  cy.url().should('match', /\/audit\/submission-progress\/[0-9]{4}-[0-9]{2}-GSAFAC-[0-9]{10}$/);
 }
 
 export function testTribalAuditPrivate(){
-    cy.get(`#is_tribal_information_authorized_to_be_public-no`).click({force:true});
-    cy.get('#tribal_authorization_certifying_official_name').type('Clint Eastwood');
-    cy.get('#tribal_authorization_certifying_official_title').type('Official');
-    cy.get('#continue').click();
-    cy.url().should('match', /\/audit\/submission-progress\/[0-9]{4}-[0-9]{2}-GSAFAC-[0-9]{10}$/);
+  cy.get(`#is_tribal_information_authorized_to_be_public-no`).click({force:true});
+  cy.get('#tribal_authorization_certifying_official_name').type('Clint Eastwood');
+  cy.get('#tribal_authorization_certifying_official_title').type('Official');
+  cy.get('#continue').click();
+  cy.url().should('match', /\/audit\/submission-progress\/[0-9]{4}-[0-9]{2}-GSAFAC-[0-9]{10}$/);
 }
